修正推流卡住的问题,修改代码后没有enable write属性导致处于拉流状态

+++ mediabase.h
@@ -586,7 +586,7 @@ public:
     int size;
     int with_adts_ = 0;
     unsigned char *data = NULL;
-    int64_t pts;
+    uint32_t pts;
 };
 
 class AudioSpecMsg :public MsgBaseObj
@@ -614,7 +614,7 @@ public:
     int type;
     int size;
     unsigned char *data = NULL;
-    int64_t pts;
+    uint32_t pts;
 };
 
 class VideoSequenceHeaderMsg :public MsgBaseObj
